requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
numpy==1.21.6
requests==2.31.0
gunicorn==22.0.0
scikit-learn==0.22.1
Pillow==7.0.0
tensorflow-gpu==2.1.0
protobuf==3.20.2
streamlit==1.23.1
streamlit-option-menu==0.3.12
